apps/devdiag-http/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
python-jose==3.3.0
requests==2.32.3
pydantic==2.9.2
mcp-devdiag[playwright,export]==0.2.1
